allowed for multiple test users and split out authentication
